--- core/model/PushBullet.ts
@@ -1,4 +1,0 @@
-import {AbstractDataObject} from './AbstractDataObject';
-
-export class PushBullet extends AbstractDataObject {
-}
