+++ src/model/AlertEmitterPushBullet.ts
@@ -1,0 +1,4 @@
+import {AbstractDataObject} from './AbstractDataObject';
+
+export class AlertEmitterPushBullet extends AbstractDataObject {
+}
